ref(nav): Move pipeline nav
